requirements.txt
openai
streamlit
audio_recorder_streamlit
streamlit-float
python-dotenv
langchain
tiktoken
pypdf==4.2.0
python-dotenv
langchain-openai
langchain-community
PyPDF2
unstructured
python-magic
unstructured[all-docs]
faiss-cpu
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-auth
langchain-google-community